requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-bootstrap5==21.1
django-filter==21.1
django-widget-tweaks==1.4.9
gunicorn==20.1.0
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
